backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pymysql
python-dotenv
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
python-multipart
PyPDF2
python-docx
pytesseract
pillow
scikit-learn
nltk
unidecode
opencv-python
numpy
requests
httpx
 